api/requirements.txt
pdfplumber
pdf2image
opencv-python
pyzbar
pillow
flask
qrcode
pytesseract
PyMuPDF
fastapi
uvicorn[standard]
requests
boto3
python-multipart
sqlalchemy